Move guest and host mappings into env.rs
