+++ sdk/src/lib.rs
@@ -5,15 +5,8 @@
 //#![feature(alloc)]
 //mod alloc;
 
-#[cfg(target_family = "wasm")]
-mod guest;
-#[cfg(target_family = "wasm")]
-pub use guest::*;
-
-#[cfg(not(target_family = "wasm"))]
-mod host;
-#[cfg(not(target_family = "wasm"))]
-pub use host::*;
+mod env;
+pub use env::*;
 
 mod object_type;
 
@@ -23,10 +16,3 @@ mod vec;
 pub use bignum::BigNum;
 pub use map::Map;
 pub use vec::Vec;
-
-#[inline(always)]
-pub const fn require(b: bool) {
-    if !b {
-        panic!();
-    }
-}
